Breaks the Styles for controls out of the Generic.xaml ResourceDictionary Removes Properties folder Updates project to Net 5 Removes custom ContextMenu Moves helper class and converters to Infrastructure folder Moves images resources to Resources folder Removes BuildProcessTemplates Replaces TestApp with BreadcrumbLib.Demo Renames demo, Model
