+++ BreadCrumbLib.Demo/Window1.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows;
+
+namespace BreadcrumbLib.Demo
+{
+	/// <summary>
+	/// Interaction logic for Window1.xaml
+	/// </summary>
+	public partial class Window1 : Window
+	{
+		public Window1()
+		{
+			InitializeComponent();
+		}
+	}
+}
