+++ BreadCrumbLib/AssemblyInfo.cs
@@ -1,0 +1,16 @@
+using System.Windows;
+using System.Windows.Markup;
+
+[assembly: ThemeInfo(
+    ResourceDictionaryLocation.None, //where theme specific resource dictionaries are located
+                                     //(used if a resource is not found in the page,
+                                     // or application resource dictionaries)
+    ResourceDictionaryLocation.SourceAssembly //where the generic resource dictionary is located
+                                              //(used if a resource is not found in the page,
+                                              // app, or any theme specific resource dictionaries)
+)]
+
+[assembly: XmlnsPrefix("http://BreadcrumbLib.com", "bc")]
+[assembly: XmlnsDefinition("http://BreadcrumbLib.com", "BreadcrumbLib")]
+[assembly: XmlnsDefinition("http://BreadcrumbLib.com", "BreadcrumbLib.Infrastructure")]
+
