+++ BreadCrumbLib/BreadCrumbItem.xaml
@@ -1,0 +1,21 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:BreadcrumbLib="clr-namespace:BreadcrumbLib">
+
+    <ResourceDictionary.MergedDictionaries>
+        <ResourceDictionary Source="pack://application:,,,/BreadcrumbLib;Component/Themes/Brushes.xaml"/>
+    </ResourceDictionary.MergedDictionaries>
+
+    <Style TargetType="{x:Type BreadcrumbLib:BreadcrumbItem}">
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="{x:Type BreadcrumbLib:BreadcrumbItem}">
+                    <BreadcrumbLib:BreadcrumbButton DataContext="{Binding RelativeSource={RelativeSource TemplatedParent}}" x:Name="button"/>
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+        <Setter Property="Focusable" Value="False"/>
+        <Setter Property="Image" Value="{Binding Path=Image, Mode=Default}"/>
+    </Style>
+
+</ResourceDictionary>
