+++ BreadCrumbLib/Breadcrumb.xaml
@@ -1,0 +1,98 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:BreadcrumbLib="clr-namespace:BreadcrumbLib">
+
+
+<ResourceDictionary.MergedDictionaries>
+		<ResourceDictionary Source="pack://application:,,,/BreadcrumbLib;Component/Themes/Brushes.xaml"/>
+	</ResourceDictionary.MergedDictionaries>
+
+    <ControlTemplate x:Key="AdditionalButton" TargetType="{x:Type Button}" >
+		<Border
+				x:Name="border"
+				Background="White"
+				BorderBrush="White"
+				BorderThickness="0,1,0,1"
+				Focusable="False"
+				SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}">
+			<StackPanel Orientation="Horizontal">
+				<Path
+						x:Name="path"
+						DockPanel.Dock="Right"
+						Stretch="Fill"
+						Stroke="#FF7F7F7F"
+						Data="m0,0 v1"
+						StrokeThickness="1"
+						Margin="0,-1.5,0,-1.5"/>
+				<ContentPresenter Focusable="False" HorizontalAlignment="Center" VerticalAlignment="Center" Margin="{TemplateBinding Padding}"/>
+			</StackPanel>
+		</Border>
+		<ControlTemplate.Triggers>
+			<Trigger Property="IsPressed" Value="True">
+				<Setter Property="Background" TargetName="border" Value="{StaticResource ButtonMouseOver}"/>
+				<Setter Property="BorderBrush" TargetName="border" Value="{StaticResource BorderPressedBrush}"/>
+				<Setter Property="BorderThickness" TargetName="border" Value="0,1,0,0"/>
+				<Setter Property="Padding" Value="0,1,0,0"/>
+			</Trigger>
+			<Trigger Property="IsMouseOver" Value="True">
+				<Setter Property="Background" TargetName="border" Value="{StaticResource ButtonMouseOver}"/>
+			</Trigger>
+		</ControlTemplate.Triggers>
+	</ControlTemplate>
+
+	<ItemsPanelTemplate x:Key="BreadcrumbPanel">
+		<StackPanel Orientation="Horizontal"/>
+	</ItemsPanelTemplate>
+
+    <Style TargetType="{x:Type BreadcrumbLib:Breadcrumb}" BasedOn="{StaticResource {x:Type ItemsControl}}">
+
+        <Setter Property="Padding" Value="1"/>
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="{x:Type BreadcrumbLib:Breadcrumb}">
+          
+                 
+                    <Border 
+                   
+                        x:Name="outerBorder" BorderBrush="#CAF3F1F4" BorderThickness="1" CornerRadius="3,3,3,3">
+                        <Border
+                            x:Name="innerBorder"
+                            BorderBrush="LightGray"
+                            BorderThickness="1"
+                            CornerRadius="2,2,2,2"
+                            Background="{x:Null}">
+                            <DockPanel>
+                                <ItemsControl
+                                    DockPanel.Dock="Right"
+                                    ItemsSource="{Binding Buttons, RelativeSource={RelativeSource TemplatedParent}}"
+                                    BorderThickness="0,0,0,0"
+                                    Focusable="False">
+                                    <ItemsControl.ItemsPanel>
+                                        <ItemsPanelTemplate>
+                                            <StackPanel Orientation="Horizontal"/>
+                                        </ItemsPanelTemplate>
+                                    </ItemsControl.ItemsPanel>
+                                    <ItemsControl.Resources>
+                                        <Style TargetType="{x:Type Button}">
+                                            <Setter Property="Template" Value="{StaticResource AdditionalButton}"/>
+                                        </Style>
+                                    </ItemsControl.Resources>
+                                </ItemsControl>
+                                <BreadcrumbLib:BreadcrumbView
+                                    x:Name="PART_View"
+                                    ItemsPanel="{StaticResource BreadcrumbPanel}"
+                                    Focusable="False"
+                                    Background="#FFFFFFFF"
+                                    BorderThickness="0,0,0,0" />
+                            </DockPanel>
+                        </Border>
+                    </Border>
+                
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+
+    </Style>
+
+
+</ResourceDictionary>
