+++ BreadCrumbLib/Infrastructure/DebugConverter.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Globalization;
-using System.Linq;
 using System.Windows.Data;
 
-namespace BreadcrumbLib
+namespace BreadcrumbLib.Infrastructure
 {
 	public class DebugConverter:IValueConverter
 	{
