+++ BreadCrumbLib/Infrastructure/ImageToImageConverter.cs
@@ -1,11 +1,10 @@
 ﻿using System;
 using System.Globalization;
-using System.Linq;
 using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
 
-namespace BreadcrumbLib
+namespace BreadcrumbLib.Infrastructure
 {
 	public class ImageToImageConverter : IValueConverter
 	{
