+++ BreadcrumbLib/Themes/Brushes.xaml
@@ -1,22 +1,23 @@
-﻿<ResourceDictionary
-	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-	<LinearGradientBrush x:Key="ButtonMouseOver" EndPoint="0.5,1" StartPoint="0.5,0">
-		<GradientStop Color="#FFFFFFFF" Offset="0"/>
-		<GradientStop Color="#FFACDEF7" Offset="1"/>
-	</LinearGradientBrush>
-	<LinearGradientBrush x:Key="GradientBorderBrush" StartPoint="0,0" EndPoint="0,1">
-		<GradientStop Color="#FF6F6E6F" Offset="0" />
-		<GradientStop Color="#FFD6D5D6" Offset="1" />
-	</LinearGradientBrush>
-	<SolidColorBrush x:Key="BorderPressedBrush" Color="#CCA1A0B6"/>
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
 
-	<SolidColorBrush x:Key="ContextMenuBackgroundBrush" Color="#FFF0F0F0" />
-	<SolidColorBrush x:Key="MenuItemHighlightBorderBrush" Color="#FF80D4F6" />
-	<LinearGradientBrush x:Key="MenuItemHighlightedBorderBackground" StartPoint="0.5,0" EndPoint="0.5,1">
-		<GradientStop Offset="0" Color="#55D6E9FC"/>
-		<GradientStop Offset="1" Color="#5596E3FF"/>
-	</LinearGradientBrush>
+    <LinearGradientBrush x:Key="ButtonMouseOver" StartPoint="0.5,0" EndPoint="0.5,1">
+        <!--<GradientStop Offset="0" Color="#FFFFFFFF" />-->
+        <GradientStop Offset="1" Color="#FFACDEF7" />
+    </LinearGradientBrush>
 
-	<!-- Resource dictionary entries should be defined here. -->
+    <LinearGradientBrush x:Key="GradientBorderBrush" StartPoint="0,0" EndPoint="0,1">
+        <!--<GradientStop Offset="0" Color="#FF6F6E6F" />-->
+        <GradientStop Offset="1" Color="#FFD6D5D6" />
+    </LinearGradientBrush>
+
+    <SolidColorBrush x:Key="BorderPressedBrush" Color="#CCA1A0B6" />
+    <SolidColorBrush x:Key="ContextMenuBackgroundBrush" Color="#FFF0F0F0" />
+    <SolidColorBrush x:Key="MenuItemHighlightBorderBrush" Color="#FF80D4F6" />
+
+    <LinearGradientBrush x:Key="MenuItemHighlightedBorderBackground" StartPoint="0.5,0" EndPoint="0.5,1">
+        <GradientStop Offset="0" Color="#55D6E9FC" />
+        <GradientStop Offset="1" Color="#5596E3FF" />
+    </LinearGradientBrush>
+
+    <!--  Resource dictionary entries should be defined here.  -->
 </ResourceDictionary>
