--- TestApp/App.xaml
@@ -1,8 +1,0 @@
-﻿<Application x:Class="TestApp.App"
-	xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	StartupUri="Window1.xaml">
-	<Application.Resources>
-		<ResourceDictionary Source="/PresentationFramework.Aero,Version=3.0.0.0,Culture=neutral,PublicKeyToken=31bf3856ad364e35,ProcessorArchitecture=MSIL;component/themes/aero.normalcolor.xaml" />
-	</Application.Resources>
-</Application>
